fix incorrect value after toggle
The active value is only updated when the counting button is clicked,
not when the toggle button is clicked. To display the most recent value
for the current counter, we somehow need to fetch this value without
triggering the active counter. The solution implemented in this commit
is to use two separate signals: an "active" signal which causes the
counter to increment when it is examined/activated, and a "passive"
signal which can be examined without being activated.