depend on gloss-algorithms, closes #13
Extent has been moved to a new package called gloss-algorithms.

As a side-effect, I had to bump the lower-bound version for our gloss
dependency. Technically the code should still compile with an older
version of gloss, but then cabal would not be able to find a version of
gloss-algorithms compatible with that version, and I don't know how to
tell cabal that gloss-algorithms would not be necessary in that case.