stack.yaml for reactive-banana-example
The examples in this repository are quite different from one another,
and I won't be surprised if they end up requiring conflicting versions
of GHC or of common dependencies. So even though they must all use my
Buttons, using a separate stack project for each example is more likely
to work than using a multi-package project for all the examples.