remove undefined, fix #39
A classic case of "this should never have worked".

The semantics of stepperD is to use the starting value until the first
event. An undefined value sometimes makes sense there, but not this
time. generatorD is used in the dynamic implementation of scenario 0:
the click counter is dynamically replaced with a new counter every time
the toggle is clicked; in particular, before the first time the toggle
is clicked, we need to start with a counter, not with undefined.