extract GlossInterface from callback-example
I want the Main file to focus on the FRP network, not on attaching the
FRP network to gloss. Some libraries already have helper functions
making it easy to start a gloss program using a network from that
library instead of the usual initial state, render, and update
functions. Let's define such a function for the libraries which don't
have one already.